+++ src/scripts/src/InstanceScripts/Instance_RazorfenDowns.cpp
@@ -44,7 +44,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -53,7 +53,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -62,7 +62,7 @@ public:
 		     spells[2].cooldown = 10;
 		     spells[2].targettype = TARGET_ATTACKING;
 		     spells[2].instant = true;
-		     spells[2].perctrigger = (float)RandomFloat(20.0f);
+		     spells[2].perctrigger = RandomFloat(20.0f);
 		     spells[2].attackstoptimer = 3000;
 		     m_spellcheck[2] = true;
 
@@ -100,7 +100,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
 
@@ -212,7 +212,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -221,7 +221,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -259,7 +259,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
 
@@ -332,7 +332,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -341,7 +341,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -379,7 +379,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
 
@@ -453,7 +453,7 @@ public:
 		     spells[0].cooldown = 10;
 		     spells[0].targettype = TARGET_ATTACKING;
 		     spells[0].instant = true;
-		     spells[0].perctrigger = (float)RandomFloat(20.0f);
+		     spells[0].perctrigger = RandomFloat(20.0f);
 		     spells[0].attackstoptimer = 3000;
 		     m_spellcheck[0] = true;
 
@@ -462,7 +462,7 @@ public:
 		     spells[1].cooldown = 10;
 		     spells[1].targettype = TARGET_ATTACKING;
 		     spells[1].instant = true;
-		     spells[1].perctrigger = (float)RandomFloat(20.0f);
+		     spells[1].perctrigger = RandomFloat(20.0f);
 		     spells[1].attackstoptimer = 3000;
 		     m_spellcheck[1] = true;
 
@@ -500,7 +500,7 @@ public:
 
     void AIUpdate()
 	{
-		float val = (float)RandomFloat(100.0f);
+		float val = RandomFloat(100.0f);
 		SpellCast(val);
     }
